feat: implemented first service contract for fetching learn progress of a single topic

+++ mynd-backend/src/main/java/de/thi/mynd/progressTracking/dto/TopicLearnProgressDto.java
@@ -12,6 +12,6 @@ public class TopicLearnProgressDto {
     public UUID topicId;
     public LearnProgressStatus status;
     public boolean completed;
-    public int percentageCompleted;
+    public double percentageCompleted;
     public List<UUID> completedContentElementIds;
 }
